Fix CI: bump pinned versions, drop non-existent packages, placeholder Agent/MCP samples
The first CI run failed because Directory.Packages.props pinned versions
(9.0.0 / 0.3.0-preview / 0.3.0-preview) that do not exist on NuGet, and
referenced packages that were never published or were merged into other
packages.

Versions bumped to actual NuGet floors as of 2026-04-29:
  Microsoft.Extensions.AI.*   9.0.0          -> 10.5.0
  Microsoft.Agents.AI.*       0.3.0-preview  -> 1.3.0
  Microsoft.Agents.AI.A2A     (added)        -> 1.3.0-preview.260423.1
  Microsoft.Agents.AI.Hosting (added)        -> 1.3.0-preview.260423.1
  ModelContextProtocol.*      0.3.0-preview  -> 1.2.0
  OllamaSharp                 4.0.0          -> 5.4.25
  OpenAI                      2.1.0          -> 2.10.0
  Anthropic.SDK               5.0.0          -> 5.10.0

Dropped (not on NuGet under that id):
  Microsoft.Extensions.AI.Ollama
  Microsoft.Extensions.AI.AzureAIInference
  Microsoft.Agents.AI.AzureAI
  ModelContextProtocol.Client (consolidated into ModelContextProtocol)
  ModelContextProtocol.Server (consolidated into ModelContextProtocol)

Microsoft.Agents.AI 1.3 and ModelContextProtocol 1.2 changed several
public APIs (ChatClientAgentOptions.Instructions/Tools, AgentThread,
GetNewThread, WorkflowBuilder ctor, Executor.From*, McpClient.OnToolsListChanged,
WithHttpServerTransport / HttpTransportType, etc.). Until samples are
rewritten against the new surface, the affected 15 samples ship a one-line
placeholder Program.cs and the original code is preserved alongside as
*.book.txt for reference. README.md documents the punch list.

Also: minor fixes to the samples that build (long token-count pattern,
split AddStandardResilienceHandler chain, drop non-installed Console
exporter), .gitignore for build outputs.

+++ samples/ch02-extensions-ai/02.4-custom-middleware/TokenBudgetChatClient.cs
@@ -6,7 +6,7 @@ public sealed class TokenBudgetChatClient(IChatClient inner, int dailyBudget) : 
 {
     private readonly Lock _gate = new();
     private DateOnly _windowStart = DateOnly.FromDateTime(DateTime.UtcNow);
-    private int _consumed;
+    private long _consumed;
 
     public override async Task<ChatResponse> GetResponseAsync(
         IEnumerable<ChatMessage> messages,
@@ -51,7 +51,7 @@ public sealed class TokenBudgetChatClient(IChatClient inner, int dailyBudget) : 
 
     private void Account(UsageDetails? usage)
     {
-        if (usage?.TotalTokenCount is not int total) return;
+        if (usage?.TotalTokenCount is not long total) return;
         lock (_gate)
         {
             _consumed += total;
